Version 4.0.9

+++ rejuv_run.py
@@ -1,7 +1,6 @@
 # Config Wizard By: Blazetamer 2013-2016
 import os
 import re
-import urllib2
 
 import downloader
 import extract
@@ -11,52 +10,42 @@ from libs import addon_able, kodi
 
 AddonTitle = kodi.addon.getAddonInfo('name')
 SiteDomain = 'TVADDONS.CO'
+dialog = xbmcgui.Dialog()
 
 wizlink = "http://indigo.tvaddons.co/wizard/updates.txt"
 cutslink = "http://indigo.tvaddons.co/wizard/shortcuts.txt"
 
 
-def JUVWIZARD():
-    filetype = 'main'
-    link = OPEN_URL(wizlink).replace('\n', '').replace('\r', '').replace('\a', '').strip()
-    path = xbmc.translatePath(os.path.join('special://home', 'addons', 'packages'))
-    url = link
-    confirm = xbmcgui.Dialog().yesno("Please Confirm",
-                                     "                Please confirm that you wish to automatically",
-                                     "            configure Kodi with all the best addons and tweaks!",
-                                     "              ", "Cancel", "Install")
-    filetype = filetype.lower()
-    if confirm:
+def JUVWIZARD(filetype='main'):
+    if xbmcgui.Dialog().yesno("Please Confirm",
+                              "                Please confirm that you wish to automatically",
+                              "            configure Kodi with all the best addons and tweaks!",
+                              "              ", "Cancel", "Install"):
+        filetype = filetype.lower()
+        if filetype == 'main':
+            addonfolder = xbmc.translatePath('special://home')
+        elif filetype == 'addon':
+            addonfolder = xbmc.translatePath(os.path.join('special://home', 'addons'))
+        else:
+            print({'filetype': filetype})
+            dialog.ok("Error!", 'filetype: "%s"' % str(filetype))
+            return
+        link = kodi.read_file(wizlink).replace('\n', '').replace('\r', '').replace('\a', '').strip()
+        # kodi.log(link)
+        if '[error]' in link:
+            print(link)
+            dialog.ok("Error!", link)
+            return
         path = xbmc.translatePath(os.path.join('special://home', 'addons', 'packages'))
-        dp = xbmcgui.DialogProgress()
-        dp.create(AddonTitle, " ", 'Downloading and Configuring ', 'Please Wait')
         lib = os.path.join(path, 'rejuv.zip')
         try:
             os.remove(lib)
         except:
             pass
         # ## ## ... ##
-        # kodi.log(url)
-        if str(url).endswith('[error]'):
-            print url
-            dialog = xbmcgui.Dialog()
-            dialog.ok("Error!", url)
-            return
-        if '[error]' in url:
-            print url
-            dialog = xbmcgui.Dialog()
-            dialog.ok("Error!", url)
-            return
-        downloader.download(url, lib, dp)
-        if filetype == 'main':
-            addonfolder = xbmc.translatePath('special://home')
-        elif filetype == 'addon':
-            addonfolder = xbmc.translatePath(os.path.join('special://home', 'addons'))
-        else:
-            print {'filetype': filetype}
-            dialog = xbmcgui.Dialog()
-            dialog.ok("Error!", 'filetype: "%s"' % str(filetype))
-            return
+        dp = xbmcgui.DialogProgress()
+        dp.create(AddonTitle, " ", 'Downloading and Configuring ', 'Please Wait')
+        downloader.download(link, lib, dp)
         xbmc.sleep(4000)
         extract.all(lib, addonfolder, dp)
         xbmc.executebuiltin("XBMC.UpdateLocalAddons()")
@@ -70,12 +59,13 @@ def JUVWIZARD():
             addon_able.set_enabled("inputstream.rtmp")
         except:
             pass
+        xbmc.executebuiltin("XBMC.UpdateLocalAddons()")
         try:
             os.remove(lib)
         except:
             pass
         if filetype == 'main':
-            link = OPEN_URL(cutslink)
+            link = kodi.read_file(cutslink)
             shorts = re.compile('shortcut="(.+?)"').findall(link)
             for shortname in shorts:
                 xEB('Skin.SetString(%s)' % shortname)
@@ -84,33 +74,9 @@ def JUVWIZARD():
                 xEB('Skin.SetBool(%s)' % enableBG16)
                 xEB('Skin.SetBool(%s)' % enableBG17)
 
-        xbmc.sleep(4000)
-        xbmc.executebuiltin('XBMC_UpdateLocalAddons()')
-        addon_able.setall_enable()
-        try:
-            addon_able.set_enabled("inputstream.adaptive")
-        except:
-            pass
-        xbmc.sleep(4000)
-        try:
-            addon_able.set_enabled("inputstream.rtmp")
-        except:
-            pass
-        # kodi.set_setting("wizardran",'true')
-
-        dialog = xbmcgui.Dialog()
+        kodi.set_setting("wizardran",'true')
         dialog.ok(AddonTitle, "Installation Complete.", "", "Click OK to exit Kodi and then restart to complete .")
         xbmc.executebuiltin('ShutDown')
-
-
-def OPEN_URL(url):
-    req = urllib2.Request(url)
-    req.add_header('User-Agent',
-                   'Mozilla/5.0 (Linux; U; Android 4.2.2; en-us; AFTB Build/JDQ39) AppleWebKit/534.30 (KHTML, like Gecko) Version/4.0 Mobile Safari/534.30')
-    response = urllib2.urlopen(req)
-    link = response.read()
-    response.close()
-    return link
 
 
 def xEB(t): xbmc.executebuiltin(t)
